feat(exports): split get* into subpath modules for tree-shaking
Move each getGroupper/getMover/getModalizer/getDeloser/getRestorer/
getObservedElement/getOutline/getCrossOrigin function into its own file
under src/get/, and expose them as package subpath exports (tabster/mover,
tabster/modalizer, etc.). src/index.ts re-exports each get* from the new
location so existing consumers that import from "tabster" keep working.

Previously all get* functions lived together in Tabster.ts, which meant
every subsystem class (GroupperAPI, MoverAPI, ModalizerAPI, DeloserAPI,
OutlineAPI, CrossOriginAPI, ObservedElementAPI, RestorerAPI) was reachable
through the same static import graph. A consumer that only needed
getModalizer still pulled the entire subsystem set into its bundle
(~60KB). With the split, a bundler can now keep only the subsystem a
consumer actually references — for example, CrossOrigin.js (~33KB) is
fully eliminated from apps that do not call getCrossOrigin.

The ESM Rollup output is switched to preserveModules so the subpath
entries resolve to physical per-module files under dist/esm/. The CJS
build is unchanged and remains a single bundled dist/index.js.

No behavior change. The runtime logic of each get* function is copied
verbatim from its original location in Tabster.ts. TabsterCore still
holds optional typed slots for each subsystem and does not import their
runtime classes.
