+++ src/get/getOutline.ts
@@ -1,0 +1,17 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { OutlineAPI } from "../Outline";
+import type * as Types from "../Types";
+
+export function getOutline(tabster: Types.Tabster): Types.OutlineAPI {
+    const tabsterCore = tabster.core;
+
+    if (!tabsterCore.outline) {
+        tabsterCore.outline = new OutlineAPI(tabsterCore);
+    }
+
+    return tabsterCore.outline;
+}
